Updated i18n and removed some console logs and fixed a few bugs.
Add some feedback for fetch since there was no way to tell if it failed.
Fixed a bug where the geometry being created through the history panel wasn't taking the crs into account.
Fixed it so that if you have a feature selected when you pull, if it encounters conflicts it unselects the feature.