Fixed a few issues.
Fixed an issue with the history panel not reloading after a commit is made.
Fixed it so that Auto-Sync will sync after the first timeout interval of 30 secs.
Made it so that if there is an empty dateTime field it will automatically set it to today's date and time.
Added a button that allows you to edit a point features geometry by setting a lat/lon position.